+++ NEWS.md
@@ -1,9 +1,12 @@
 # RFishBC 0.0.3 ongoing
 * `digitizeRadii()`: Modified. Slight modification to the messages in the console. Re worked the code with `locator()` so that the points will be shown "snapped to the transect" if `snap2Transect=TRUE` (this addresses [#7](https://github.com/droglenc/RFishBC/issues/7)). Added an `orig.pts` data.frame to the RData object which contains the original (non-snapped to transect) points (which can be plotted with `showDigitizedImage()`). Changed `addTransect=` to `showTransect=` to make more similar to `showDigitezedImages()`.
-* `showDigitizedImage()`: Modified. Added `showOrigPts=`, `pch.show2=`, `col.show2=`, and `cex.show2=` to handle including original points on the image.
+* `RFBCoptions()`: Modified. Added `showAnnuliLabels=`, `col.ann=`, and `cex.ann=` for use in `showDigitizedImage()`. Changed defaults of mostly colors, pchs, and cexs.
+* `showDigitizedImage()`: Modified. Added `showAnnuliLabels=`, `col.ann=`, and `cex.ann=` to shown annuli numbers when just one transect is shown. Added `showOrigPts=`, `pch.show2=`, `col.show2=`, and `cex.show2=` to handle including original points on the image.
 * `iFindTransect()`: Added. Moved this code out of `iSelectAnnuli()`.
 * `iGetImage()`: Modified. Streamlined code. Added `native=TRUE` to `read.bitmap()` call to send to the underlying functions as this is apparently more efficient when using `rasterImage()` (which this uses).
+* `iOrderPts()`: Added. Made it easier to add annuli labels in `showDigitizedImage()`.
 * `iSelectAnnuli()`: Modified. Now calls `iSelectTransect()`.
+* `iShowAnnuliLabels()`: Added. Made it easier to add annuli labels in `showDigitizedImage()`.
 * `ishowTransect()`: Added. Moved this code out of `iSelectAnnuli()`. Also called from `showDigitizedImage()`.
 
 # RFishBC 0.0.2 27-Apr-18